Move BlockchainBalances to their own file
